♻️ [#93] Refactor: callback 로직을 api와 hook으로 분리
